+++ calamity/Calamity/Types/Model/Channel/Webhook.hs
@@ -3,6 +3,7 @@
 -- | Channel webhooks
 module Calamity.Types.Model.Channel.Webhook (Webhook (..)) where
 
+import Calamity.Internal.Redaction (redactToken)
 import {-# SOURCE #-} Calamity.Types.Model.Channel
 import {-# SOURCE #-} Calamity.Types.Model.Guild.Guild
 import Calamity.Types.Model.User
@@ -11,7 +12,7 @@ import Data.Aeson ((.:), (.:?))
 import Data.Aeson qualified as Aeson
 import Data.Text (Text)
 import Optics.TH
-import TextShow.TH
+import TextShow qualified
 
 data Webhook = Webhook
   { id :: Snowflake Webhook
@@ -23,8 +24,30 @@ data Webhook = Webhook
   , avatar :: Text
   , token :: Maybe Text
   }
-  deriving (Eq, Show)
+  deriving (Eq)
   deriving (HasID Webhook) via HasIDField "id" Webhook
+
+instance Show Webhook where
+  showsPrec _ Webhook {id, type_, guildID, channelID, user, name, avatar, token} =
+    showString "Webhook {id = "
+      . shows id
+      . showString ", type_ = "
+      . shows type_
+      . showString ", guildID = "
+      . shows guildID
+      . showString ", channelID = "
+      . shows channelID
+      . showString ", user = "
+      . shows user
+      . showString ", name = "
+      . shows name
+      . showString ", avatar = "
+      . shows avatar
+      . showString ", token = "
+      . shows (redactToken <$> token)
+      . showChar '}'
+
+deriving via TextShow.FromStringShow Webhook instance TextShow.TextShow Webhook
 
 instance Aeson.FromJSON Webhook where
   parseJSON = Aeson.withObject "Webhook" $ \v -> do
@@ -41,5 +64,4 @@ instance Aeson.FromJSON Webhook where
       <*> v .: "avatar"
       <*> v .:? "token"
 
-$(deriveTextShow ''Webhook)
 $(makeFieldLabelsNoPrefix ''Webhook)